android/app/src/main/java/id/ac/ugm/sv/trpl/glucosemonitoring/domain/usecase/GetProfileUseCase.kt
package id.ac.ugm.sv.trpl.glucosemonitoring.domain.usecase

import id.ac.ugm.sv.trpl.glucosemonitoring.domain.model.User
import id.ac.ugm.sv.trpl.glucosemonitoring.domain.repository.IUserRepository
import io.reactivex.rxjava3.android.schedulers.AndroidSchedulers
import io.reactivex.rxjava3.core.Flowable
import io.reactivex.rxjava3.schedulers.Schedulers
import javax.inject.Inject

class GetProfileUseCase @Inject constructor(
    private val userRepository: IUserRepository,
) {
    
    operator fun invoke(): Flowable<User> {
        return userRepository.getProfile()
            .subscribeOn(Schedulers.io())
            .observeOn(AndroidSchedulers.mainThread())
    }
    
}